requirements.txt
litellm==1.77.4
python-dotenv==1.1.1
rich==14.1.0
